requirements.txt
Click==7.0
Flask==1.0.3
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.3.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.2
SQLAlchemy==1.3.4
Werkzeug==0.15.4
xmltodict==0.12.0
